어댑터 refactor
